requirements.txt
torch
torchvision
opencv-python
prettytable
click
tensorboard
tqdm
pillow
matplotlib
